New branch with UUID h9lcq8ep9i66ppfwklez1dma2
Initial branch nickname: log_lifetime_writes

Initial check-out location: /etc

Initial version

Provide a libexec script (which can also be used as a cron job) for
recording the total number of bytes ever written to some media volume.

It will typically be used for flash-based storage to monitor the amount
of data written to it, so that this amount is known when the medium is
finally worn out.

Then it is possibly to verify the "TBW"-claims of the drive manufacturer
against actual data, or calculate the write amplification overhead from
this.

The script itself contains the name of the block device to monitor as
its last name component; rename to script (or create symlinks for it) in
order to monitor arbitrary block devices.

Every individually-named instance of the script (such as symlinks to it)
maintains its own data file in /var/lib where it updates the write
counters every time it is called.

Those status files are human-readable and contain the gathered sums of
writes to that device. They are *not* log files, they are just updated
with new values by the script. Which means they will not grow in size
themselves (other than getting a few bytes larger as the increasing
counts require more digits for representation).

The number of writes recorded by the script will typically not be exact
but rather a good approximate.

The problem here is that the script reads kernel-provided counters which
will be reset after every reboot.

If your box crashes before the script has a chance to record the latest
value of those counters before rebooting, the bytes written since the
last run of the script will be lost and are not included in the counts.

For the same reason, not only a cron job but also the shutdown scripts
for regular shutdowns should invoke the scripts as late as possible
before actually unmounting the volume containing /var/lib, thus
including the latest possible value for the write counter in the sums
saved in the status files.

Regarding the frequency for running the cron jobs: The more often they
run, the less unrecorded write counter increment will be lost in case of
a crash.

But if it is run too often, updating the status files themselves may
shorten the life of the flash drive.

On a heavy-loaded system which permanently writes data to the flash
drive, adding a few bytes of additionally written data will not make any
practical difference.

For such cases, running the cron job every hour seems to be a good idea.

Otherwise, for long-running systems which are idle most of the time, I
recommend running the crob job daily.

Generally, the less data is written to the drive between the invocation
of the cron job, the less often it needs to be called.

For systems which write really little (typically primarily RAM-disk
based ones which only occasionally write persistent user data to the
flash drive), even running the cron job once a week may still be enough.