Remote cron job verifier for lifetime writes
Comparison of the two cron jobs led to the conclusion that the "lifetime
writes" cron-job generally does the right thing.

The negative deltas in size which raised concerns seem to be the effect
of reboots without properly saving the old state, which is outside of
the scope a cron job can do something about.